Rename JRSwizzle files to LPSwizzle to avoid conflicts in projects where JRSwizzle is used
